*: Reformat cmtadm
cmtadm is well along the way to be reformatted too.
To aid in this the definition of FilePath has been modified;
this adds the joinpath() method to FilePath, which means we
can avoid converting back and forth between PurePath and FilePath,
and thus shorten quite a lot of lines. Begin by converting cmtadm;
other files to follow.

Signed-off-by: David Weinehall <david.weinehall@intel.com>