[WIP] Work on VM image support
Support for template-based installation of VM-based
worker nodes is *almost* working...  Things are a little
rough around the edges still, an example template file
is needed, some improved errors too, but the main issue
is that while the template image creation works properly,
that template doesn't seem to be usable as a backing file
for other images.

If we cannot prepare one single image and clone it
we're going to have major performance issues. Doing
potentially hundreds of separate VM installations
just isn't feasible.  Sadly it's tricky to debug.

Still, we better at least create a git branch with this.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>