[WIP] Keep trying to get things working
The same problem remains; non-functioning network
for the instances.  It seems that the template image we use
as backing image somehow gets tied to the MAC address we use
during the creation of that image, so even if we force
cloud-init to be re-run it still uses that MAC.

netplan *claims* that it should be regenerating its config
on every boot, so you'd hope that this means that it should
handle changing MAC-address. Whether this is true or not
is probably our best lead when it comes to fixing this though,
so that's next on the TODO list.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>