*: Version-check improvements
Generalise version checks a bit.  While we still have a ways
to go we at least have a hopefully more resilient way to check
for the Kubernetes upstream version.  Use the same method to
check for the latest version of crc.  While crc does show this
information too we cannot rely on the proxy settings being
correct when crc checks.

Add a version entry for crc to __VersionData.yaml.

The next step is to call updated_version_cache() from
__VersionData.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>