networkio.py,*: Fix get_github_version()
get_github_version() checked for the latest release
chronologically, not the latest release lexically,
meaning that if a new patch revision for an earlier
minor release is released after the latest minor release,
it'll be thought to be the latest release. Thus we might
end up missing important fixes.  This would obviously be bad.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>