[WIP]: Work to make CMT packagable
Work on making it possible to install CMT to the system path
(and, thanks to use of DESTDIR, also creating Debian-packages).
With this commit most things should work, with the following caveats:

* Most/all tests are broken at the moment.
* The configuration files and data files are not yet read from
  the system path.

It will still be possible to override views, parser-files,
themes, sources, and playbooks, as well as, of course, cmt.yaml
from the home directory, but by default the data should come
from the system path.

We also need to handle the first time a new user runs CMT;
since we won't be running cmt-install anymore, we can no longer
install ~/.cmt that way; we need to create it, and various
other when necessary.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>