*: Add template support for views
Since the old and the new Event API use the exact same
event reasons (hopefully this won't change) and since we
use those reasons twice in each file, *aaand* since there
are at the moment almost 350 reasons, it's no longer feasible
to keep these files in sync manually.

There will probably be other uses for this further down
the line; for instance it would probably make sense to
use templates for common shortcuts.

This adds a build-dependency on python3-jinja2.
It is currently not used run-time.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>