Use 0 as default for missing timestamp
seconds_to_age takes an int, not a string, so we shouldn't
use <unset> as default if the value is unset.

Ideally we should handle this better, but a policy needs
to be defined how. Once we have proper logging in place
we could probably log an error on type mismatch and use
a sensible default, but until then an exception is still
better than silently accepting mismatches that might be signs
of programming errors.

Reported-by: Eero Tamminen <eero.t.tamminen@intel.com>
Signed-off-by: David Weinehall <david.weinehall@intel.com>