requirements.txt: pin dependency on setuptools
There's a theoretical (documented as requiring
a social engineering attack) issue in setuptools,
which is depending on by ansible-runner through
one layer of indirection.

Rather than have scanners complain we might as well
bump the dependency. While at it depend on the latest
patch revision of ansible-runner in the 2.1-series.

We should start depending on a newer minor verson
of ansible-runner though, it seems that there are
no API-breakage.

Signed-off-by: David Weinehall <david.weinehall@intel.com>