Don't allow upgrades with modified feature gates
(By default) we shouldn't allow upgrading the control plane
when the set of feature gates have been modified, since it may
fail and leave the cluster in an unrecoverable state.

Provide an option to override it, but default to aborting
installation if changes are detected.

Since we're adding a new commandline option we need to
regenerate the helptexts and add bash completion.
Add a missing bash completion while at it.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>