logparser.py: Try to improve bare key handling
Some logs contains combinations of both lines with key=value
messages and lines with other formats. With the introduction
of support for bare keys we opened a can of worms that could
cause lines such as "foo bar baz" be treated as a key value
line. Theoretically it can be, but it's most likely not.

Let's require that there's at least one "real" key=value
pair on the line to parse it as a key=value log entry,
and check that the first pair key on the line doesn't end
with a ":" (which typically signals a facility).

Signed-off-by: David Weinehall <david.weinehall@gmail.com>