views/__Container: Ellipsise list of pods
There can be a *lot* of pods using the same container image.
Rather than trying to list all of the pods that uses it,
ellipsise after 5 pods. The info view can still show all the pods
(in a vertical manner), but listing them all horizontally
eventually causes curses to crash.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>