Makefile: Ignore semgrep false positive
One of the semgrep tests for jinja2 seems to assume that
Jinja2 is only used for HTML and flags use of Jinja2 as
a risk for cross site scripting. Seeing as the use in CMT
is for templating YAML-files, since it's only used during the
build process, and since the input variables that are used
for templating are not supplied externally, we can safely
ignore this warning.

Signed-off-by: David Weinehall <david.weinehall@intel.com>