Lots of typing and misc fixes
With these fixes the typing scores are good enough for release.
Not only were plenty of type issues fixed, the transition to
Python 3.9+ style type signatures also made some headway.

In most cases it improves legibility.

The transition should of course continue, but it's lower
priority than getting the test coverage up to releasable
levels; currently we're below the coverage for v0.8.4.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>