Include mdtable.py and use it to format data
To simplify the preparation of the changelog for a new release,
use mdtable.py to reformat data to something Markdown friendly.
Theoretically we don't need to pad the data (Markdown is clever
enough to handle lines of different width as long as they
use the separators), but when the lines vary in length it
becomes impossible to read in code view.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>