cmu,views: Fix a crash and some overlapping views
The time comparison tried using total_seconds instead of
int(total_seconds()). 2 errors in one statement. Not bad...

Also, some of the Kind + API-families we support only
differ in family; the Kind is not unique. So we cannot
have both use the same default_command. Prioritise built-ins,
and use aliases for the rest.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>