Type-fixes, docstrings, and a bit of cleanup
Add a lot of type signatures. Fix some issues found by mypy.
Add a lot of docstrings. Rewrite some functions with optional
args to use kwargs. Make mypy and pylint a bit stricter
and fix some of their findings.

Some of the rewrites may have unintended consequences
(regressions); the changes were been made to make it easier
to refactor the code by cutting down on the amount of type
overloading. logparser.py, genericinfoloop(),
genericlistloop(), containerinfoloop(), and clusteroverview()
are all too complex to allow this though.

Due to the risk of regressions and the difficulty of testing
this in an automated manner, the program will undergo
manual testing for the next few days (while the missing
docstrings are added) before releasing v0.8.5.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>