Start fixing yamllint complaints
So, it turns out that the CI system had a .yamllint file
laying around that ignored a lot of warnings.

We're talking > 600 files, so fixing this will take a while
(manual fixes only, to ensure that things don't get unreadable
 in the process).

The same process will also have to be repeated on parser-files,
themes, playbooks, and any other yaml-files we might have.

Still, it feels worthwhile to do before release; previous
release-notes have (incorrectly) claimed that we're yamllint-clean.
This time we should live up to that claim.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>