views,*: New views
Add several new views related to Rancher.
Add a new feature to the list generator; the ability
to specify different formatter for each field in a list
item; currently only the default formatter and numerical with units
is supported. Next we should probably add address.
If possible to merge this with the status-version of generator_list
that would be a bonus.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>