validatortests: Disable two non-working testcases
Two of the testcases report different results depending
on whether the test is called directly on the command line
or whether it's called as part of coverage-testing.

This is obviously not acceptable, so we disable
the testcases for now. The coverage will be slightly worse,
but the code hasn't changed for a long while,
and the only difference a failure on this code path
would make is what error message is printed, so it's
a safe thing to do.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>