Make it possible to delete an object and its owner
Add a flag (delete_owner) to delete_object() that allows
for deleting an object's owner at the same time as the object.

The only user, for now, of this functionality is Kueue Workloads.
They behave in a bit magical manner. Workloads are (normally)
not created directly. Instead they are created when a job
is created with some extra annotations.

So if the user tries to delete a Workload it'll just be recreated.
This provides nice functionality (it allows for requeueing
a Workload. But it also means that to really delete a workload
you first need to track down the corresponding Job, delete that Job,
then delete the Workload.

With this functionality available behind the scenes there's now
a new action available in the Workloads view:
Delete resource (and controller);
the old Delete resource command has been renamed Requeue resource.

Note: actionfunc() seems to take extra arguments in several superfluous
ways. This should be simplified; they should only take actionfunc_args,
similar to infogetter_args, listgetter_args, etc.

Signed-off-by: David Weinehall <david.weinehall@intel.com>