requirements.txt: Bump jinja2 dependency
Dependabot reported that the required version
of jinja2 has two security issues. There is no known
path to exploiting them in cmt, since the only way
templates can be fed to cmt is by the user,
who already has whatever permissions cmt runs with.

But since others might copy code from cmt and use it
in a manner when an exploit would lead to privilege
escalation, we should still fix it.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>