Fix align_and_pad() behaviour
Fix calling convention for align_and_pad() to match
the behaviour. We treated the pad-argument as a bool,
meaning that only the pre-defined pad-length, or no pad at all,
was applied. Change the pad argument to a bool to enforce
this in practice. While at it call the function with
named arguments, use default values, and introduce TypedDicts
for the formatting dict.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>