Implement a new version of cmtlog
The old version of cmtlog was dropped a long time ago,
always with the intention of replacing it with a solution
using Python's standard logging module but using a custom
format. This implementation is exactly that.

From now on we can add logging for all recoverable errors,
security violations, etc.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>