*: Add views for KRO Instances
Add views for KRO Instances. This doesn't fully solve the issue
of the instances being dynamically created and thus tricky
to fit into cmu's current way of functioning, but at least
it makes them possible to view.

Ideally we need a way to populate CRDs dynamically
with template views for "known unknowns" as well as
"unknown unknowns".

Signed-off-by: David Weinehall <david.weinehall@intel.com>