setup_cri-o: Fix the cri-o playbook
Switch to another repo. This will most not work behind proxies,
but currently installing doesn't work at all.  The new repo doesn't
have support for cri-o < 1.28, so the old repo is still used for those
(but seeing as the signing key has expired it's unlikely to work
unless the key is fixed).

Signed-off-by: David Weinehall <david.weinehall@gmail.com>