cmu,*: Improved error handling & notifications
Simplify the code for the notification & alert boxes a bit.
Replace some cases of sys.exit() with alert() + aborting the failed
action. In general we should strive to alert + abort + log,
ignore + log, or in special cases raise + log rather than
use sys.exit().

In some places of the code validation is done before use
rather than when populating the data. A fail fast solution
would be preferable. Some, but not all, of those cases
have been tagged in the code.

Signed-off-by: David Weinehall <david.weinehall@intel.com>