Further logging improvements
The viewer for the logs needs to handle the logs nicely.
The logpad in the infoview doesn't quite meet the standards
of the logpad for containers though, this clearly needs
to be unified, log handling simplified, etc.

Sigh.

Also, things got shuffled around a bit, so several
testcases will fail. But it's only a matter of moving them
to the right place to get them working again.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>