cni_data,cmtadm: Improve cilium upgrade
The regex for cilium CNI was incorrect, leading to cilium
never detecting new versions (only the binary was upgraded).
Once that was fixed another issue unearthed itself;
one of the arguments to upgrade was incorrectly used
as a global argument instead.

To help debug such issues in the future cmtadm now
prints the upgrade command when using --verbose.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>