Prepare for splitting configmap signatures
We'll want to split configmap type signatures from the code,
to make it easier to add new signatures without having to
modify the code. This will simplify adding local modifications
that are not suitable for upstreaming.

This file is not used by anything yet, so remember to keep
its content in sync with any changes made to formatters.py
until the latter is rewritten to use this file.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>