playbooks: Fix CNI issues
At least on some distros there seems to be some confusion
as to whether CNI plugins should be read from /opt/cni/bin
or /usr/lib/cni.

Ideally both should be supported, since distro-provided
files should be in a sub-directory of /usr/lib to adhere
to FHS. This isn't a path that CNI plugins should modify though,
so it's /opt/cni/bin is a suitable place for them to put it.

Containerd (at least) seems to have an issue with this, however,
and there doesn't seem to be any way to configure it to search
both paths, so for the time being we rely on the upstream
kubernetes-cni package (which stores plugins in /opt/cni/bin)
instead of the distro-supplied containernetworking-plugins package,
and configure containerd to get plugins from /opt/cni/bin.

Hopefully this can be reverted at some point.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>