Add playbook aliases
For the benefit of cmtinv (or rather for the benefits
of the users thereof), let's define some aliases for
the various playbooks. It might not be easy to memorise
the exact name of a playbook (is is update_os, upgrade_os,
update-os, or upgrade-os?), so list several variants.

This is just preparatory work; cmtinv still needs to
be modified to handle the aliases.

Signed-off-by: David Weinehall <david.weinehall@gmail.com>